changed mpg type due to missing data in CSV

+++ db/migrate/20140827203610_create_cars.rb
@@ -4,8 +4,8 @@ class CreateCars < ActiveRecord::Migration
       t.string :make
       t.string :model
       t.integer :year
-      t.float :city_mpg
-      t.float :highway_mpg
+      t.integer :city_mpg
+      t.integer :highway_mpg
 
       t.timestamps
     end
